==> study1_novel/step2_fiction_booknlp.ipynb <==
import pandas as pd
from ast import literal_eval
def return_liter_eval(x):
    try:
        return list(literal_eval(x))
    except ValueError:
        return None

import numpy as np

meta_df = pd.read_csv("./gutenberg_meta_df.csv")

print(meta_df.shape)
meta_df.head()

import os
from tqdm import tqdm
from IPython.display import clear_output

books = meta_df.id.unique()

from glob import glob

len(glob("../../gutenberg_standard/new_analysis/proc_text_new/*_text.txt"))

import seaborn as sns

from booknlp.booknlp import BookNLP
model_params={
                "pipeline":"entity,quote,supersense,event,coref", 
                "model":"small"
                }

booknlp=BookNLP("en", model_params)

books.shape

# these files are failed for NLP analysis for too large files, 
# to avoid break the loop, 
# skip these books
skip = ["PG17530", "PG3254", "PG4500", "PG5600", "PG5400"]

len(glob("../../gutenberg_standard/new_analysis/book_nlp_output_small/*/*.book"))


# file_size = []
for b in tqdm(books):
    inputFile = "../../gutenberg_standard/new_analysis/proc_text_new/{}_text.txt".format(b)
    outputDir = "../../gutenberg_standard/new_analysis/book_nlp_output_small/{}/".format(b)
    idd = b
    
    if not os.path.exists(outputDir):
        os.mkdir(outputDir)

    if (not os.path.exists(outputDir + "{}.book".format(b))) and (b not in skip):
        print("Processing book: {} ->> file size: {} mb".format(b, (os.path.getsize(inputFile)/ (1024 * 1024))))
        booknlp.process(inputFile, outputDir, idd)
        
        clear_output(wait=True)

book_nlp_outputs = glob("../../gutenberg_standard/new_analysis/book_nlp_output_small/*")

len(book_nlp_outputs)
